requirements.txt
pytest
requests
python-dotenv
curlify
allure-pytest
Faker